+++ galactory/__init__.py
@@ -2,26 +2,15 @@
 # (c) 2022 Brian Scholer (@briantist)
 
 import logging
-import typing as t
 
-from datetime import datetime
 from flask import Flask, request
-from flask.json.provider import DefaultJSONProvider
 from configargparse import ArgParser, ArgumentError, Action
 from artifactory import ArtifactoryPath
 
 from .api import bp as api
 from .download import bp as download
 from .health import bp as health
-
-
-class DateTimeIsoFormatJSONProvider(DefaultJSONProvider):
-    @staticmethod
-    def default(o: t.Any) -> t.Any:
-        if isinstance(o, datetime):
-            return o.isoformat()
-
-        return super().default(o)
+from .utilities import DateTimeIsoFormatJSONProvider
 
 
 def create_app(**config):
